Move GUI code to own namespace
