+++ RedBookPlayer/App.xaml.cs
@@ -5,6 +5,7 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
+using RedBookPlayer.GUI;
 
 namespace RedBookPlayer
 {
